+++ desktop/cmp/panel/Panel.js
@@ -152,8 +152,9 @@ function parseLoadDecorator(prop, name, contextModel) {
     if (isValidElement(prop))               return prop;
     if (prop === 'onLoad') {
         const loadModel = contextModel?.loadModel;
+        console.log(contextModel);
         if (!loadModel) {
-            console.warn(`Cannot use 'onLoad' for '${name}'.  Context model does not have LoadSupport or have a 'loadModel' property.`);
+            console.warn(`Cannot use 'onLoad' for '${name}'.  Context model does not implement loading.`);
             return null;
         }
         return cmp({model: loadModel, spinner: true});
@@ -189,7 +190,6 @@ Panel.propTypes = {
      *   + true for a default LoadingIndicator,
      *   + a PendingTaskModel for a default LoadingIndicator bound to a pending task,
      *   + the string 'onLoad' for a default LoadingIndicator bound to the loading of the current model.
-     *     (current model must include LoadSupport).
      */
     loadingIndicator: PT.oneOfType([PT.instanceOf(PendingTaskModel), PT.element, PT.bool, PT.string]),
 
@@ -199,7 +199,6 @@ Panel.propTypes = {
      *   + true for a default mask,
      *   + a PendingTaskModel for a default load mask bound to a pending task,
      *   + the string 'onLoad' for a default load mask bound to the loading of the current model.
-     *     (current model must include LoadSupport).
      */
     mask: PT.oneOfType([PT.instanceOf(PendingTaskModel), PT.element, PT.bool, PT.string]),